Revert fastavro to non-optional dependency
Currently, skipping the avro extras would lead to an ImportError.

It doesn't really make sense to make fastavro optional in the context of
deserialization, no way to know ahead of time that the messages being
deserialized definitely won't contain Avro schemas.

This could only work if the deserializer had a plan for what to do when
encountering a message serialized with a schema that requires a skipped
dependency (raise an exception? skip the message? call a callback?)